requirements.txt
bayesian-optimization
catboost
einops
filelock
graphviz
hiddenlayer
lightgbm
matplotlib
nbconvert
notebook
numpy
ofa
pandas
pillow
pymoo~=0.5.0
pysot~=0.2.3
pytorch-forecasting
pyyaml
scikit-image
scikit-learn
scipy
seaborn
setuptools
shap
statsmodels
timm
torch
torchvision
tornado
torchprofile
tqdm



